pyproject.toml
[tool.poetry]
name = "mantra"
version = "0.1.0"
description = "MANTRA: Multi-view ANalysis with Tensor and matRix Alignment"
authors = ["Kevin De Azevedo <deazevedo@med.uni-frankfurt.de>", "Yusuf Berk Oruc", "Florian Buettner <florian.buettner@dkfz-heidelberg.de>"]
readme = "README.md"
packages = [{include = "mantra"}]
license = "MIT"
keywords = ["tensor-decomposition", "bayesian-inference", "multi-omics", "factor-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

[tool.poetry.dependencies]
python = "^3.11"
torch = ">=2.0.0, !=2.0.1"
pyro-ppl = "^1.8.6"
numpy = "^1.25.1"
pandas = ">=2.0.0"
scipy = "^1.14.1"
mudata = "^0.2.3"
scanpy = "^1.9.8"
scikit-learn = "^1.4.0"
tqdm = ">=4.60.0"

[tool.poetry.group.dev.dependencies]
seaborn = ">=0.12.2"
matplotlib = ">=3.7.0"
graphviz = "^0.20.1"
jupyterlab = "^4.0.4"
tensorly = "^0.8.1"

[tool.poetry.group.analysis.dependencies]
hdbscan = "^0.8.33"
leidenalg = "^0.10.2"
pynndescent = "^0.5.11"
gseapy = "^1.1.5"
scikit-misc = "^0.5.1"
mygene = "^3.2.2"

[tool.poetry.group.lint.dependencies]
black = "^24.4.2"
ruff = "^0.4.4"
mypy = "^1.10.0"
pre-commit = "^3.7.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
extend-select = ["B", "C4", "I", "N", "PIE", "RUF", "SIM", "TCH", "UP"]
# Ignore naming rules for mathematical notation (A1, A2, A3, R, X, Y) and PyTorch F convention
ignore = ["N803", "N806", "N812"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
files = ["mantra"]

[tool.pytest.ini_options]
testpaths = ["mantra/tests", "tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["mantra"]
omit = ["mantra/tests/*", "tests/*"]
